[Algo] Delete file and modify sort in progress
