+++ algorithm/sort.py
@@ -1,0 +1,11 @@
+def bubble_sort(nums):
+    for i in range(len(nums)):
+        for j in range(len(nums)-1):
+            if nums[j] > nums[j+1]:
+                temp = nums[j]
+                nums[j] = nums[j+1]
+                nums[j+1] = temp
+    return nums
+
+nums = [2, 1, 5, 3, 6, 2, 5]
+print(bubble_sort(nums))
